+++ src/main/java/net/helix/hlx/conf/HelixConfig.java
@@ -8,7 +8,7 @@ import java.io.File;
  *  A container for all possible configuration parameters of SBX.
  *  In charge of how we parse the configuration from given inputs.
  */
-public interface HelixConfig extends APIConfig, NodeConfig, HXIConfig, DbConfig, ConsensusConfig, ZMQConfig, TipSelConfig, PoWConfig, SolidificationConfig, GraphConfig, LoggingConfig {
+public interface HelixConfig extends APIConfig, NodeConfig, HXIConfig, DbConfig, ConsensusConfig, ZMQConfig, TipSelConfig, PoWConfig, SolidificationConfig, GraphConfig, LoggingConfig, SpamConfig {
     File CONFIG_FILE = new File("hlx.ini");
     /**
      * Parses the args to populate the configuration object