+++ changelog.md
@@ -13,6 +13,10 @@
         - MIN_TRANSACTION_AGE_THRESHOLD set to 5 seconds
         - MAX_TRANSACTION_AGE_THRESHOLD set to 2 minutes
     - Fix: Timestamp conversion in Node / API
+    - Fix SignedFiles: line seperator should not be part of digest
+    - New Resource files
+    - Added configurable spammer for testing/experiments 
+    
 # 0.5.3
     - PoW Integration (Replace divepearler with GreedyMiner)
     - PoW: difficulty is a number of zero bytes instead of a power of 2